app/requirements.txt
requests
uvicorn
fastapi
opentelemetry-distro
opentelemetry-exporter-otlp
